requirements.txt
psycopg[binary]==3.2.9
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.2
xgboost==2.1.1
fastapi==0.115.6
uvicorn==0.32.1
